fixed 5,6,3

+++ 0x08-recursion/6-is_prime_number.c
@@ -11,7 +11,7 @@
  */
 int is_prime_number(int n)
 {
-	if ((n % 2.0) != 0)
+	if (((double)n % 2.0) != 0)
 	{
 		return (1);
